requirements.txt
# Core
numpy==2.1.3
pandas==2.2.2
tqdm

# PDF & document ingestion
pdfplumber
pdf2image
pytesseract
Pillow

# Embeddings & vector search
sentence-transformers
faiss-cpu

# App / config / LLM
streamlit
python-dotenv
google-generativeai

